create package for generator and move files
Preparation for new eventGenerator extension
