+++ EapWebInterface/src/main/java/de/hpi/unicorn/application/pages/input/generator/GeneratePanel.properties
