+++ EapWebInterface/src/main/java/de/hpi/unicorn/application/pages/input/generator/GeneratorPage.html
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE HTML>
+<html xmlns:wicket="http://git-wip-us.apache.org/repos/asf/wicket/repo?p=wicket.git;a=blob_plain;f=wicket-core/src/main/resources/META-INF/wicket-1.5.xsd;hb=master"
+      xmlns="http://www.w3.org/1999/xhtml">
+<wicket:extend>
+    <div wicket:id="tabs" class="tabpanel"></div>
+</wicket:extend>
+</html>
