+++ EapWebInterface/src/main/java/de/hpi/unicorn/application/pages/input/replayer/ReplayerPage.java
@@ -57,11 +57,6 @@ public class ReplayerPage extends AbstractEapPage {
 			}
 		});
 
-		tabs.add(new AbstractTab(new Model<String>("Generate events")) {
-			@Override
-			public Panel getPanel(final String panelId) { return new GeneratePanel(panelId, page); }
-		});
-
 		this.add(new BootstrapTabbedPanel<ITab>("tabs", tabs));
 		setOutputMarkupId(true);
